requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pillow
psycopg2-binary
alembic
python-multipart
requests
minio
# Programas Sistema
# inkscape 
# RClone